fix: bottom nav bg back, create material, handle iOS swipe back gesture dynamically

+++ src/features/farm/components/pondwork/measurement/MeasurementDataBox.tsx
@@ -105,7 +105,7 @@ export const MeasurementDataBox: React.FC<MeasurementDataBoxProps> = ({
                     onShrimpSizeChange(formatDecimalInput(text, 15, 5));
                 }}
                 keyboardType="numeric"
-                inputFormat={InputFormat.DECIMAL}
+                inputFormat={InputFormat.INTEGER}
                 required
                 maxLength={21}
                 containerStyle={{ marginBottom: 0 }}
